+++ webui/test/JobView.test.js
@@ -18,195 +18,168 @@
  */
 import test from 'ava';
 import Vue from 'vue';
-import JobView from '../components/jobs/JobsView';
+import JobView from '../components/jobs/detail/JobView';
+import { METRIC_LIST } from '../components/jobs/detail/JobView';
 
 test.before('initialize', async t => {
   // load all plugins
   require('../plugins/event-bus');
 });
 
-test.serial('selectJobId', async t => {
+test.serial('beforeMount', async t => {
   const vm = new Vue(JobView).$mount();
-  let counter = 0;
-  let dagEmitted = null;
 
-  vm.$eventBus.$on('job-id-select', async (data) => {
-    counter++;
-  });
-
-  vm.$eventBus.$on('dag', async (data) => {
-    dagEmitted = data.dag;
-  });
-
-  vm._newJob('foo');
-  vm._newJob('bar');
-  vm.selectedJobId = 'foo';
-  vm.jobs.foo.dag = 'lorem';
-
-  await vm.selectJobId('bar');
-  t.is(counter, 1, 'job-id-select event should be fired');
-  t.is(vm.selectedJobId, 'bar', 'selectedJobId should be changed');
-  t.is(dagEmitted, null, 'dag should not be changed');
-
-  await vm.selectJobId('bar');
-  t.is(counter, 1, 'job-id-select event should not be fired');
-  t.is(vm.selectedJobId, 'bar', 'selectedJobId should not be changed');
-  t.is(dagEmitted, null, 'dag should not be changed');
-
-  await vm.selectJobId('foo');
-  t.is(counter, 2, 'job-id-select event should be fired');
-  t.is(vm.selectedJobId, 'foo', 'selectedJobId should be changed');
-  t.is(dagEmitted, 'lorem', 'dag should be changed');
-
-  vm.$eventBus.$off('job-id-select');
-  vm.$eventBus.$off('dag');
-  vm.$destroy();
-});
-
-test('deletejobId', async t => {
-  const vm = new Vue(JobView).$mount();
-  let deselectCounter = 0;
-  let clearDagCounter = 0;
-
-  vm.$eventBus.$on('job-id-deselect', async (data) => {
-    deselectCounter++;
-  });
-
-  vm.$eventBus.$on('clear-dag', async (data) => {
-    clearDagCounter++;
-  });
-
-  vm._newJob('foo');
-  vm._newJob('bar');
-  vm.selectedJobId = 'foo';
-
-  vm.deleteJobId('foo');
-  await vm.$nextTick();
-
-  // should be deleted
-  t.falsy('foo' in vm.jobs,
-    'job `foo` should be deleted from `jobs` object');
-  t.is(deselectCounter, 1, 'job-id-deselect event should be fired');
-  t.is(clearDagCounter, 1, 'clear-dag event should be fired');
-  t.is(vm.selectedJobId, '', 'selectedJobId should be cleared');
-
-  vm.selectedJobId = 'foo';
-  vm.deleteJobId('bar');
-  await vm.$nextTick();
-
-  // if selectedJobId and target id is not same, do not emit clear events
-  // also, make sure that job id is not changed
-  t.falsy('bar' in vm.jobs,
-    'job `bar` should be deleted from `jobs` object');
-  t.is(deselectCounter, 1,
-    'job-id-deselect event should not be fired when selectedJobId is differ from deleted job id');
-  t.is(clearDagCounter, 1,
-    'clear-dag event should not be fired when selectedJobId is differ from deleted job id');
-  t.is(vm.selectedJobId, 'foo', 'job id should not be changed');
-
-  vm.$eventBus.$off('job-id-deselect');
-  vm.$eventBus.$off('clear-dag');
-  vm.$destroy();
-});
-
-test('processMetric', async t => {
-  const vm = new Vue(JobView).$mount();
-  let toBeCalledCounter = 0;
-  vm.processIndividualMetric = async () => {
-    toBeCalledCounter++;
-  };
-
-  // individual metric
-  await vm.processMetric({metricType: 'foo'}, 'foo');
-  t.is(toBeCalledCounter, 1, 'individual metric should be processed');
-
-  // big chunk of initial metric
-  const fakeMetric = {
-    FooMetric: {
-      FooID: {
-        metricType: 'FooMetric',
-        data: {}
-      },
-      BarID: {
-        metricType: 'FooMetric',
-        data: {}
-      },
-      BazID: {
-        metricType: 'FooMetric',
-        data: {}
-      }
-    },
-    BarMetric: {
-      LoremID: {
-        metricType: 'BarMetric',
-        data: {}
-      },
-      IpsumID: {
-        metricType: 'BarMetric',
-        data: {}
-      }
-    }
-  };
-
-  await vm.processMetric(fakeMetric, 'foo');
-  t.is(toBeCalledCounter, 6, 'metric chunk should be properly processed');
+  t.is(vm.groupDataSet.length, METRIC_LIST.length,
+    'groupDataSet should be built with METRIC_LIST');
 
   vm.$destroy();
 });
 
-test.serial('processIndividualMetric', async t => {
+test.serial('job id event handler', async t => {
   const vm = new Vue(JobView).$mount();
-  let stateChangeEventData = null;
-  let dagData = null;
+  let setTimelineItemsCounter = 0;
+  let clearStageSelectCounter = 0;
 
-  vm.$eventBus.$on('state-change-event', data => {
-    if (stateChangeEventData) {
-      t.fail('state-change-event event should be fired only once');
-    }
-    stateChangeEventData = data;
+  vm.$eventBus.$on('set-timeline-items', data => {
+    setTimelineItemsCounter++;
   });
 
-  vm.$eventBus.$on('dag', data => {
-    if (dagData) {
-      t.fail('dag event should be fired only once');
-    }
-    dagData = data;
+  vm.$eventBus.$emit('job-id-select', {
+    jobId: 'foo',
+    jobFrom: 'bar',
+    metricLookupMap: 'lorem',
+    metricDataSet: 'ipsum',
+  });
+  await vm.$nextTick();
+
+  t.is(vm.selectedJobId, 'foo', 'job id should be changed correctly');
+  t.is(vm.jobFrom, 'bar', 'jobFrom should be changed correctly');
+  t.is(vm.metricLookupMap, 'lorem',
+    'metricLookupMap should be changed correctly');
+
+  vm.$eventBus.$emit('job-id-deselect');
+  await vm.$nextTick();
+
+  t.is(setTimelineItemsCounter, 1,
+    'set-timeline-items event should be fired again');
+  t.is(vm.selectedJobId, '', 'job id should be cleared');
+  t.is(vm.jobFrom, '', 'jobFrom should be cleared');
+  t.deepEqual(vm.metricLookupMap, {}, 'metricLookupMap should be cleared');
+
+  vm.$destroy();
+});
+
+test.serial('table data event handler', async t => {
+  const vm = new Vue(JobView).$mount();
+  let tableData = '';
+
+  vm.buildTableData = (data) => { tableData = data; };
+
+  vm.selectedMetricId = 'lorem';
+  vm.selectedJobId = 'ipsum';
+
+  vm.$eventBus.$emit('build-table-data', {
+    metricId: 'foo',
+    jobId: 'bar',
+  });
+  await vm.$nextTick();
+
+  t.is(tableData, '',
+    'buildTableData should not be called id is different');
+
+  vm.$eventBus.$emit('build-table-data', {
+    metricId: 'lorem',
+    jobId: 'ipsum',
+  });
+  await vm.$nextTick();
+
+  t.is(tableData, 'lorem', 'buildTableData should be called');
+
+  vm.$destroy();
+});
+
+test.serial('metric selection event handler', async t => {
+  const vm = new Vue(JobView).$mount();
+  let metricSelectDoneCounter = 0;
+  let metricDeselectDoneCounter = 0;
+
+  vm.$eventBus.$on('metric-select-done', () => {
+    metricSelectDoneCounter++;
   });
 
-  vm._newJob('foo');
-  const job = vm.jobs.foo;
-  t.truthy(job, 'job should be valid');
+  vm.$eventBus.$on('metric-deselect-done', () => {
+    metricDeselectDoneCounter++;
+  });
 
-  const fakeMetric = {
-    metricType: 'JobMetric',
-    data: {
-      id: 'bar',
-      dag: 'bar-dag',
-      stateTransitionEvents: [
-        {
-          prevState: 'READY',
-          newState: 'EXECUTING',
-          timestamp: +new Date()
-        }
-      ],
+  vm.buildTableData = (data) => {
+    t.is(data, 'foo', 'metric-select should be deliver correct metric id');
+  };
+
+  vm.$eventBus.$emit('metric-select', 'foo');
+  await vm.$nextTick();
+
+  t.is(vm.selectedMetricId, 'foo',
+    'metric-select event should change selected metric id');
+  t.is(metricSelectDoneCounter, 1,
+    'metric-select event should fire metric-select-done event');
+
+  vm.$eventBus.$emit('metric-deselect');
+  await vm.$nextTick();
+
+  t.deepEqual(vm.tableData, [],
+    'metric-deselect event should clear tableData');
+  t.is(vm.selectedMetricId, '',
+    'metric-deselect event should clear selectedMetricId');
+  t.is(metricDeselectDoneCounter, 1,
+    'metric-deselect event should fire metric-deselect-done event');
+
+  vm.$eventBus.$off('metric-select-done');
+  vm.$eventBus.$off('metric-deselect-done');
+  vm.$destroy();
+});
+
+test.serial('buildTableData functionality', t => {
+  const vm = new Vue(JobView).$mount();
+
+  vm.metricLookupMap = {
+    fooMetric: {
+      group: 'dog',
+      content: 'cat',
+
+      executionProperties: {
+        epFoo: 'epFooValue',
+      },
+
+      fooProperty: -1,
+      barProperty: 'ruff ruff',
     }
   };
 
-  vm.buildMetricLookupMapWithDAG = (jobId) => { t.is(jobId, 'foo', 'job id should be matched'); };
-  await vm.processIndividualMetric(fakeMetric, 'foo');
-  await vm.$nextTick();
+  vm.buildTableData('fooMetric');
+  t.is(3, vm.tableData.length);
 
-  t.is(stateChangeEventData.jobId, 'foo', 'data should be matched');
-  t.is(stateChangeEventData.metricId, 'bar', 'data should be matched');
-  t.is(stateChangeEventData.metricType, 'JobMetric', 'data should be matched');
-  t.is(stateChangeEventData.prevState, 'READY', 'data should be matched');
-  t.is(stateChangeEventData.newState, 'EXECUTING', 'data should be matched');
+  for (const item of vm.tableData) {
+    switch (item.key) {
+      case 'executionProperties':
+        t.truthy('extra' in item,
+          'extra field should valid in executionProperties data');
+        t.is(item.extra[0].key, 'epFoo',
+          'executionProperties should be built correctly');
+        t.is(item.extra[0].value, 'epFooValue',
+          'executionProperties should be built correctly');
+        break;
+      case 'fooProperty':
+        t.is(item.value, 'N/A', '-1 should be converted to N/A');
+        break;
+      case 'barProperty':
+        t.is(item.value, 'ruff ruff', 'other values should be kept');
+        break;
+      case 'group':
+      case 'content':
+        t.fail('`group` or `content` property should be discarded');
+        break;
+    }
+  }
 
-  t.is(dagData.dag, 'bar-dag', 'data should be matched');
-  t.is(dagData.jobId, 'foo', 'data should be matched');
-  t.false(dagData.init, 'data should be matched');
-
-  vm.$eventBus.$off('state-change-event');
-  vm.$eventBus.$off('dag');
   vm.$destroy();
 });
